Group code by feature (#614)
Why:
Currently we have a few poorly organized features. By grouping feature logic in the same place we make the code base easier to reason about and maintain over time.

This commit:
Is purely a refactor, no new functionality has been added.

Groups the product analytics logic by feature
Renames analytics to ProductAnalytics to better match the domain concept and differentiate from the CovidData feature.
Groups the self assessment logic by feature
Moves device util from gaen module to a Device module
Moves non-gaen device info out of gaen module to the Device module
Removes unused keyboard avoiding view
