+++ src/AffectedUserFlow/CodeInput/EnableExposureNotifications.spec.tsx
@@ -1,11 +1,12 @@
 import React from "react"
-import { openAppSettings } from "../../gaen/nativeModule"
 import { render, fireEvent } from "@testing-library/react-native"
 
+import { openAppSettings } from "../../Device"
 import EnableExposureNotifications from "./EnableExposureNotifications"
 
-jest.mock("../../gaen/nativeModule")
+jest.mock("../../Device/nativeModule")
 jest.mock("@react-navigation/native")
+
 describe("EnableExposureNotifications", () => {
   it("displays the explanations for the exposure notifications", () => {
     const { getByText } = render(<EnableExposureNotifications />)
@@ -16,6 +17,7 @@ describe("EnableExposureNotifications", () => {
     ).toBeDefined()
     expect(getByText("Enable exposure notifications to continue")).toBeDefined()
   })
+
   it("takes user to system settings if decide to activate notifications", () => {
     const { getByLabelText } = render(<EnableExposureNotifications />)
