+++ src/AffectedUserFlow/CodeInput/EnableExposureNotifications.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "react-i18next"
 
 import { Text, StatusBar } from "../../components"
 import { useStatusBarEffect } from "../../navigation"
-import { openAppSettings } from "../../gaen/nativeModule"
+import { openAppSettings } from "../../Device"
 
 import { Spacing, Colors, Typography, Buttons } from "../../styles"
