+++ src/Device/index.ts
@@ -1,0 +1,23 @@
+import {
+  doesDeviceSupportLocationlessScanning,
+  getApplicationName,
+  getBuildNumber,
+  getVersion,
+  isBluetoothEnabled,
+  isLocationEnabled,
+  openAppSettings,
+  subscribeToBluetoothStatusEvents,
+  subscribeToLocationStatusEvents,
+} from "./nativeModule"
+
+export {
+  doesDeviceSupportLocationlessScanning,
+  getApplicationName,
+  getBuildNumber,
+  getVersion,
+  isBluetoothEnabled,
+  isLocationEnabled,
+  openAppSettings,
+  subscribeToBluetoothStatusEvents,
+  subscribeToLocationStatusEvents,
+}
