+++ src/SelfAssessment/EmergencySymptomsQuestions.tsx
@@ -9,7 +9,7 @@ import { SelfAssessmentStackScreens, useStatusBarEffect } from "../navigation"
 import { Text } from "../components"
 
 import { EmergencySymptom, SymptomGroup } from "./selfAssessment"
-import { useSelfAssessmentContext } from "../SelfAssessmentContext"
+import { useSelfAssessmentContext } from "./Context"
 import SymptomCheckbox from "./SymptomCheckbox"
 import SelfAssessmentLayout from "./SelfAssessmentLayout"
