--- src/healthAuthority.ts
@@ -1,3 +1,0 @@
-export interface HealthAuthority {
-  name: string
-}
