+++ src/navigation/MainNavigator.tsx
@@ -12,9 +12,10 @@ import {
 } from "@react-navigation/native"
 import { useTranslation } from "react-i18next"
 
-import { ModalStackScreens, HomeStackScreens } from "./index"
 import { useOnboardingContext } from "../OnboardingContext"
-import { useAnalyticsContext } from "../AnalyticsContext"
+import { useAnalyticsContext } from "../ProductAnalytics/Context"
+
+import { ModalStackScreens, HomeStackScreens } from "./index"
 import { WelcomeStackScreens, Stacks } from "./index"
 import MainTabNavigator from "./MainTabNavigator"
 import HowItWorksStack from "./HowItWorksStack"
@@ -25,7 +26,7 @@ import AgeVerification from "../modals/AgeVerification"
 import LanguageSelection from "../modals/LanguageSelection"
 import ProtectPrivacy from "../modals/ProtectPrivacy"
 import AffectedUserStack from "../AffectedUserFlow/"
-import AnonymizedDataConsentScreen from "../modals/AnonymizedDataConsentScreen"
+import AnonymizedDataConsentScreen from "../ProductAnalytics/AnonymizedDataConsentScreen"
 import SelfAssessmentStack from "./SelfAssessmentStack"
 import ExposureDetectionStatus from "../Home/ExposureDetectionStatus"
 import BluetoothInfo from "../Home/BluetoothInfo"
@@ -33,6 +34,7 @@ import ExposureNotificationsInfo from "../Home/ExposureNotificationsInfo"
 import LocationInfo from "../Home/LocationInfo"
 import CallbackStack from "./CallbackStack"
 import CovidDataDashboard from "../CovidData/Dashboard"
+
 import { applyModalHeader } from "./ModalHeader"
 import { applyHeaderLeftBackButton } from "../navigation/HeaderLeftBackButton"
